+++ merlin/models/tf/models/ranking.py
@@ -11,8 +11,8 @@ from merlin.models.tf.core.combinators import ParallelBlock, TabularBlock
 from merlin.models.tf.inputs.base import InputBlockV2
 from merlin.models.tf.inputs.embedding import EmbeddingOptions, Embeddings
 from merlin.models.tf.models.base import Model
-from merlin.models.tf.models.utils import parse_prediction_tasks
 from merlin.models.tf.prediction_tasks.base import ParallelPredictionBlock, PredictionTask
+from merlin.models.tf.prediction_tasks.utils import parse_prediction_tasks
 from merlin.models.tf.transforms.features import CategoryEncoding
 from merlin.schema import Schema, Tags
